Tweaks to modal dropdown picker
